backend/requirements.txt
chromadb
uvicorn
langchain
langchain-community
langchain-groq
langchain-chroma
python-dotenv
fastapi[all]